requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.12.0
faiss-cpu>=1.7.4
numpy>=1.24.0

